requirements.txt
jira
pyyaml
click
keyring
Flask
gunicorn
markdown2
pytz
openpyxl
Flask-SQLAlchemy
Flask-Login
Werkzeug
cryptography
